fixes cyclic dependency due to libp2p utils
