+++ network/p2p/node/libp2pNode.go
@@ -17,6 +17,7 @@ import (
 	"github.com/libp2p/go-libp2p/core/protocol"
 	"github.com/libp2p/go-libp2p/core/routing"
 	"github.com/onflow/flow-go/network/p2p"
+	"github.com/onflow/flow-go/network/p2p/internal/p2putils"
 	"github.com/rs/zerolog"
 
 	"github.com/onflow/flow-go/network/slashing"
@@ -173,7 +174,7 @@ func (n *Node) CreateStream(ctx context.Context, peerID peer.ID) (libp2pnet.Stre
 
 // GetIPPort returns the IP and Port the libp2p node is listening on.
 func (n *Node) GetIPPort() (string, string, error) {
-	return IPPortFromMultiAddress(n.host.Network().ListenAddresses()...)
+	return p2putils.IPPortFromMultiAddress(n.host.Network().ListenAddresses()...)
 }
 
 func (n *Node) RoutingTable() *kbucket.RoutingTable {
